Kill webob requirement and move expirer to proxy
This turned out to be more involved. The object-expirer had wrong
config file. In RHOS we got away with adding a pipeline to the
object server. But as it turns out the proper way to do it is to
supply it with its own config. The daemon then loads it twice:
as a daemon and when InternalClient is instantiated. Fine. But
how create instance'd config for SAIO? There is no special directory
like /etc/object-server into which to drop instances of expirer.conf.
For now we nix that option (it is still better than the nonworkable
config we shipped up to now).

You cannot yet get rid of python-webob on your nodes, because for
some unfathomable reason, python-keystone requires it, and we require
python-keystone even on the nodes. This needs fixing next time around.