Add the /var/cache/swift bz#870409
Not having the recon directory is generally harmless, but
1. it produces tracebacks into logs
2. someone is likely to use recon and will create it anyway

I don't see a downside and ownding directories in /var/cache
appears to be permitted.

Since recon cache is overwritten on each dump, we could locate
this in /var/run (/run), but I'm not quite sure if it's worth
the bother and surprise. No, I don't know exactly why Debian
chose /var/cache. Possibly deceived by the word "cache" in the
phrase "recon cache".

See bz#870409.