Drop python-swiftclient as a dependency
We actually had a sad experience with following upstream on this,
when Derek had to re-install dependency to python-keystoneclient
for the sake of painless Packstack installs. So, caveat emptor.

The main risk is some kind of tool, which does import swiftclient.
Hopefuly Ubuntu plumbed that cesspool before us. The code fix was
in upstream for a little while now.

See bz#1058131 for defails.