Redirect logging to /var/log/swift/swift.log
The thorniest problem with making Swift log like any other OpenStack
component is that it uses syslog. Therefore, it must use one of the
pre-defined logging facilities, and we cannot be sure that there is
no conflict. We're taking this risk because the current situation is
worse. See the bug 997983 for the long deliberation.

Resolves: rhbz#997983