Make fast-post default
Now that container-sync works with fast-post, we can make it
default. The upstream keeps post-as-copy default, for compatibility
reasons, but the new code should work fine. So, this patch amounts
to the following:

  [filter:copy]
  use = egg:swift#copy
  object_post_as_copy = true

However, if we add the "copy" filter, we might as well add
all the others that we traiditionaly use, to give it a correct slot.
And if we do that, we might as well provide sample pipelines.

This patch carries a certain risk:

 - Scripts may process our configurations and may choke at the
   changes

 - There may be a bug in fast-post, container-sync as it relates
   to the fast-post, or elsewhere

 - We're reducing the defauls and so more explicit configuration
   may be needed in the future (although Director and Puppet
   should be configuring for us anyway).

But plunging into the world of fast-post seems worthy of it.

Related: rhbz#1334547
Change-Id: I476cb2d80e6c5e34d941b2044305cff84159aef7