Update patch set 3
Patch Set 3: Code-Review-1

I fully agree with David; Swift itself does not depend on the ceilometermiddleware and works fine without it. With this change any standalone deployment has to install the ceilometermiddleware (and any additional dependencies that come with it).

I rather prefer to fix this in puppet-swift, because puppet-swift configures the ceilometermiddleware to be in the proxy pipeline (IIRC).

We're already using a similar approach with the swift3 middleware: Swift itself doesn't need this, but it is configured in puppet-swift, therefore the dependency will be also installed by puppet-swift. I suggest to use the same solution in this case.

Patch-set: 3