Update patch set 3
Patch Set 3: Code-Review+1

I'm +1 on this, +2 if we add a comment to explain that this is a temp. workaround to fix an issue found in TripleO.

About the permanent fixsolution, I had a quick look at the issue this morning, and it looks like the swift-proxy package is pre-installed in the overcloud image, so we'd need fix the image to include ceilometermiddleware. Everything else should be covered by the existing patches to puppet-swift.

Patch-set: 3
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1