Update patch set 3
Patch Set 3:

Do you guys think it would be useful if composite configuration for Swift also assembled the proxy pipeline automatically? I think it does not currently. That would allow Swift to start early and be useful to the rest of the deployment, but without Ceilometer middleware, to be installed later if OOO (and Puppet modules) wish do.

I realize Pradeep's point about not pulling extra dependencies because pre-exising Keystone middleware does, but if we ever find a way not to pull that one in, we'd still have Puppet assume that all is well. It's very difficult to undo this kind of debt.

Patch-set: 3
Reviewer: Gerrit User 76 <76@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0