Permission: Match what tripleo does after install.
There is currently a difference between the permission the packages
install files with and the permissions applied by the puppet manifest.

This difference results in rpm -Va being noisy when run, which in turns
result in certification of the platform to fail. This commit aims to
sync permissions between packages and what puppet does.

Change-Id: I71dc1a694aad3de7be78911d16a32e9715f998b5