Update patch set 1
Patch Set 1:

I'd like to see this cleaned up, but I have some small concerns.

The default upstream is to use LOCAL0:

http://paste.openstack.org/show/607117/

And on a clean TripleO deployment it looks like this:

http://paste.openstack.org/show/607115/

So we need at least some sanity checks on puppet-swift as well to clean this up. We also need to ensure to set LOCAL2 as the new default for all services in puppet-swift. There is a good chance for some regression in the future, isn't it?

Maybe it is simpler to clean up the LOCAL1/LOCAL2, set them to LOCAL0 in Swift (falling back to default) and change the haproxy logging then?

Patch-set: 1