Update patch set 1
Patch Set 1:

I changed the proxy setting here: https://review.openstack.org/#/c/456387/ so at least the LOCAL1 isn't present anymore in the undercloud. The reconcilier is not exposed in puppet-swift yet, so it would need a patch there...

What do you think is the best solution? I don't think using local0 is a good practice, as it overrides with other services. For example I can also see Keepalived_healthcheckers in local0 on the undercloud.

Patch-set: 1