Add HUP to syslogd post swift package installation
Swift by default logs to syslog.  The swift package already caters to
redirecting this to the correct location of /var/log/swift/swift.log,
but due to there being no existing reload of the syslog daemon, this
change is ineffective.  By default as it stands, the undercloud has no
logging to /var/log/swift/swift.log until the operator sends `kill -1`
to syslogd, after which logs flow correctly.

This patch corrects this behaviour via a post install HUP to syslogd.

Change-Id: I4d91ee63392dc2ace68c1fb2782d6828a102ca72
Closes-Bug: 1646451