Update patch set 1
Patch Set 1:

> syslog isn't under systemctl control

I don't know what you mean by the "not under systemctl control". Everything in the system is under systemd's control, it's process 1 and launches everything too - using cgroups so that daemons cannot escape its control. It is not safe to assume that PID file is even going to be present in the system. Well, to be sure, it's not safe to assume that rsyslogd is present either, since you can replace it with journald now. Maybe that systemctl reload rsyslod also needs the || : action.

Patch-set: 1