Add missingok to logrotate
The logrotate was observed in the field to run before Swift
itself had a chance to serve any requests, and thus no log
file was yet created. In such a case, logrotate sends a nag
e-mail to the administrator, which is undesirable: there is
no problem, it is purely a distracting noise.

Change-Id: I755554797f047eb69051efef5da17e980a8d6b68
Resolves: rhbz#1536061