Update patch set 6
Patch Set 6: Workflow-1

We may have a problem with TripleO jobs, since they do not pre-install openstack-macros, and it would be a requirement for swift in this case. Let's not merge it for now, until we can think of a good solution.

Patch-set: 6
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=-1