Swift to the distributed object expirer
We should have done this for Ussuri, but fortunately the deprecation
cycles in Swift are very long, especially for something this invasive.
But it's time to get on with it.

The essense of the change is that object expirers run on object
server nodes now. So much effort was spent on explaining to
packagers that object-expirer service belongs to the
openstack-swift-proxy package, and now we basically have
to revert to what used to be a packaging mistake.

We set dequeue_from_legacy to true, because there's no
alternative at present.

This commit also rearranges some configs and other files
so they end in a proper subpackage.

Change-Id: Ia5385098d3b47a365978bba42a6e09d603913cec