Swift to the distributed object expirer
We should have done this for Ussuri, but fortunately the deprecation
cycles in Swift are very long, especially for something this invasive.
But it's time to get on with it.

The essense of the change is that object expirers run on object
server nodes now. So much effort was spent on explaining to
packagers that object-expirer service belongs to the
openstack-swift-proxy package, and now we basically have
to revert to what used to be a packaging mistake.

A thorny problem with this is, do we set dequeue_from_legacy
to true or not. In the new world, operators are supposed to
choose a node (such as a proxy), where to run the legacy
expirer with that option. We just don't have a way to express
this kind of asymmetry in RPMs unless we create several new
packages (for the old expirer and the new expirer separately).
For now we just let all the new expirers to scan the legacy
account and contend for it. It's the right thing to do for
small clusters, and can be tuned in big clusters.

This commit also rearranges some configs and other files
so they end in a proper subpackage.

Change-Id: Ia5385098d3b47a365978bba42a6e09d603913cec