Update patch set 3
Patch Set 3:

> or maybe create a symlink object-expirer.conf pointint to object-server.conf?

I'm not sure it's going to work, because the object-expirer.conf had a pipeline for the internal client, and the new expirer code checks for the name "expirer" in the config. If no such name found, it assumes that it uses object-server.conf and looks for the pipeline in the internal-client.conf. If you symlink like that, the expirer will attempt to run with a full pipline of the proxy, I expect. Not sure what's going to happen then. Well, all of the middlewares in the pipeline are configured...

Patch-set: 3